--- Project2/Debug/Project2.log
@@ -1,2 +1,0 @@
-﻿  GraphDriver.cpp
-  Project2.vcxproj -> C:\Users\Wyatt\Google Drive\College Stuff\Semester 4\COSC 3020\projects\Project 2\Project2\Debug\Project2.exe
