--- Project2/Debug/Project2.tlog/Project2.lastbuildstate
@@ -1,2 +1,0 @@
-#TargetFrameworkVersion=v4.0:PlatformToolSet=v140:EnableManagedIncrementalBuild=false:VCToolArchitecture=Native32Bit:WindowsTargetPlatformVersion=8.1
-Debug|Win32|C:\Users\Wyatt\Google Drive\College Stuff\Semester 4\COSC 3020\projects\Project 2\Project2\|
