--- Project2/test.txt
@@ -1,11 +1,0 @@
-10
-0 1 -1
-1 2 -1
-2 3 -1
-3 4 -1
-4 5 -1
-5 6 -1
-6 7 -1
-7 8 -1
-8 9 -1
-9 3 6 8 9 -1
