--- Project2/test1.txt
@@ -1,6 +1,0 @@
-4
-0	1 2 -1
-1	3 -1
-2	-1
-3	-1
-
