--- Project2/test2.txt
@@ -1,9 +1,0 @@
-8
-0	1 2 -1
-1	3 4 5 -1
-2	-1
-3	-1
-4	-1
-5	6 -1
-6	7 -1
-7	-1
