--- Project2/test3.txt
@@ -1,8 +1,0 @@
-7
-0	1 -1
-1	2 3 -1
-2	4 -1
-3	5 -1
-4	-1
-5	-1
-6	-1
