--- Project2/test4.txt
@@ -1,11 +1,0 @@
-10
-0 -1
-1 -1
-2 -1
-3 -1
-4 -1
-5 -1
-6 7 -1
-7 8 -1
-8 9 -1
-9 -1
