--- Project2/x64/Debug/Project2.log
@@ -1,2 +1,0 @@
-﻿  GraphDriver.cpp
-  Project2.vcxproj -> C:\Users\Wyatt\Desktop\Project2\x64\Debug\Project2.exe
